Day 28
Problem : Check if a Binary Tree is Symmetric
You are given the root of a binary tree. Your task is to determine whether the tree is symmetric. A binary tree is symmetric if the left and right subtrees are mirror images of each other.

Input :
* The root of the binary tree.
* Input : [1, 2, 2, 3, 4, 4, 3]

Output :
Return true if the binary tree is symmetric, otherwise return false.
Output: true